+++ src/twg2/text/stringSearch/StringSearchOp.java
@@ -1,0 +1,28 @@
+package twg2.text.stringSearch;
+
+/**
+ * @author TeamworkGuy2
+ * @since 2016-08-07
+ */
+public enum StringSearchOp {
+	EXACT() {
+		@Override
+		public boolean test(String a, String b) { return a != null ? a.equals(b) : b == null; }
+	},
+	CONTAINS() {
+		@Override
+		public boolean test(String a, String b) { return a != null ? a.contains(b) : b == null; }
+	},
+	STARTS_WITH() {
+		@Override
+		public boolean test(String a, String b) { return a != null ? a.startsWith(b) : b == null; }
+	},
+	ENDS_WITH() {
+		@Override
+		public boolean test(String a, String b) { return a != null ? a.endsWith(b) : b == null; }
+	};
+
+
+	public abstract boolean test(String a, String b);
+
+}